- adjusted my submitted code to work with latest compiler changes
R=r
OCL=14734
CL=14734

+++ usr/gri/gosrc/import.go
@@ -41,7 +41,7 @@ func (I *Importer) ReadByte() byte {
 
 func (I *Importer) ReadInt() int {
 	x := 0;
-	s := 0;  // TODO eventually Go will require this to be a uint!
+	s := uint(0);
 	b := I.ReadByte();
 	for b < 128 {
 		x |= int(b) << s;
